move the stories from profile reducer to story reducer
