+++ frontend/src/reducers/profile_reducer.js
@@ -1,4 +1,4 @@
-import { RECEIVE_PROFILE, RECEIVE_PROFILE_ERRORS, RECEIVE_STORIES } from '../actions/profile_actions';
+import { RECEIVE_PROFILE, RECEIVE_PROFILE_ERRORS } from '../actions/profile_actions';
 
 // TODO: handle the response properly
 const profileReducer = (state = {}, action) => {
@@ -7,9 +7,6 @@ const profileReducer = (state = {}, action) => {
             return action.profile ;
         case RECEIVE_PROFILE_ERRORS:
             return action;
-        case RECEIVE_STORIES:
-            const stories = action.stories.map(o => { return { title: o.title, body: o.body } });
-            return Object.assign({}, { stories: stories });
         default:
             return state;
     }
